+++ kvs/include/huaweicloud/kvs/v1/model/GetKvRequestBody.h
@@ -13,9 +13,7 @@
 #include <huaweicloud/core/http/HttpResponse.h>
 
 #include <string>
-#include <huaweicloud/kvs/v1/model/Projection_blob.h>
 #include <huaweicloud/core/bson/Bson.h>
-#include <vector>
 
 namespace HuaweiCloud {
 namespace Sdk {
@@ -62,34 +60,12 @@ public:
     void unsetprimaryKey();
     void setPrimaryKey(const Document& value);
 
-    /// <summary>
-    /// 对kv_doc有效，返回哪些字段列表，默认全部。 &gt; 字段名或嵌套字段路径用&#39;/&#39;分割。
-    /// </summary>
-
-    std::vector<std::string>& getProjectionFields();
-    bool projectionFieldsIsSet() const;
-    void unsetprojectionFields();
-    void setProjectionFields(const std::vector<std::string>& value);
-
-    /// <summary>
-    /// 
-    /// </summary>
-
-    Projection_blob getProjectionBlob() const;
-    bool projectionBlobIsSet() const;
-    void unsetprojectionBlob();
-    void setProjectionBlob(const Projection_blob& value);
-
 
 protected:
     std::string tableName_;
     bool tableNameIsSet_;
     Document primaryKey_;
     bool primaryKeyIsSet_;
-    std::vector<std::string> projectionFields_;
-    bool projectionFieldsIsSet_;
-    Projection_blob projectionBlob_;
-    bool projectionBlobIsSet_;
 
 };