+++ kvs/include/huaweicloud/kvs/v1/model/ScanKvRequestBody.h
@@ -14,9 +14,7 @@
 
 #include <huaweicloud/kvs/v1/model/Condition_expression.h>
 #include <string>
-#include <huaweicloud/kvs/v1/model/Projection_blob.h>
 #include <huaweicloud/core/bson/Bson.h>
-#include <vector>
 
 namespace HuaweiCloud {
 namespace Sdk {
@@ -99,24 +97,6 @@ public:
     void unsetfilterExpression();
     void setFilterExpression(const Condition_expression& value);
 
-    /// <summary>
-    /// 对kv_doc有效，返回哪些字段列表，默认全部。
-    /// </summary>
-
-    std::vector<std::string>& getProjectionFields();
-    bool projectionFieldsIsSet() const;
-    void unsetprojectionFields();
-    void setProjectionFields(const std::vector<std::string>& value);
-
-    /// <summary>
-    /// 
-    /// </summary>
-
-    Projection_blob getProjectionBlob() const;
-    bool projectionBlobIsSet() const;
-    void unsetprojectionBlob();
-    void setProjectionBlob(const Projection_blob& value);
-
 
 protected:
     std::string tableName_;
@@ -131,10 +111,6 @@ protected:
     bool endKeyIsSet_;
     Condition_expression filterExpression_;
     bool filterExpressionIsSet_;
-    std::vector<std::string> projectionFields_;
-    bool projectionFieldsIsSet_;
-    Projection_blob projectionBlob_;
-    bool projectionBlobIsSet_;
 
 };