+++ kvs/include/huaweicloud/kvs/v1/model/ScanKvResponse.h
@@ -44,7 +44,7 @@ public:
     /// ScanKvResponse members
 
     /// <summary>
-    /// 返回的文档数量，为_0_不表示结束， &gt; 如果filtered_count超过_500_仍无匹配，则返回_0
+    /// 返回的文档数量，为0不表示结束， &gt; 如果filtered_count超过500仍无匹配，则返回0
     /// </summary>
 
     int32_t getReturnedCount() const;