+++ kvs/include/huaweicloud/kvs/v1/model/ScanSkeyKvRequestBody.h
@@ -14,9 +14,7 @@
 
 #include <huaweicloud/kvs/v1/model/Condition_expression.h>
 #include <string>
-#include <huaweicloud/kvs/v1/model/Projection_blob.h>
 #include <huaweicloud/core/bson/Bson.h>
-#include <vector>
 
 namespace HuaweiCloud {
 namespace Sdk {
@@ -64,7 +62,7 @@ public:
     void setHintIndexName(const std::string& value);
 
     /// <summary>
-    /// 数据量不超过_1_mb时，返回的文档个数，最大_100_个，默认_1_mb_或者_100_个文档。
+    /// 数据量不超过1mb时，返回的文档个数，最大100个，默认1mb或者100个文档。
     /// </summary>
 
     int32_t getLimit() const;
@@ -108,24 +106,6 @@ public:
     void unsetfilterExpression();
     void setFilterExpression(const Condition_expression& value);
 
-    /// <summary>
-    /// 对kv_doc有效，返回哪些字段列表，默认全部。 - 数组元素的字段名或嵌套字段路径用&#39;/&#39;分割，e_name为下标。
-    /// </summary>
-
-    std::vector<std::string>& getProjectionFields();
-    bool projectionFieldsIsSet() const;
-    void unsetprojectionFields();
-    void setProjectionFields(const std::vector<std::string>& value);
-
-    /// <summary>
-    /// 
-    /// </summary>
-
-    Projection_blob getProjectionBlob() const;
-    bool projectionBlobIsSet() const;
-    void unsetprojectionBlob();
-    void setProjectionBlob(const Projection_blob& value);
-
 
 protected:
     std::string tableName_;
@@ -142,10 +122,6 @@ protected:
     bool endSortKeyIsSet_;
     Condition_expression filterExpression_;
     bool filterExpressionIsSet_;
-    std::vector<std::string> projectionFields_;
-    bool projectionFieldsIsSet_;
-    Projection_blob projectionBlob_;
-    bool projectionBlobIsSet_;
 
 };