+++ kvs/include/huaweicloud/kvs/v1/model/ScanSkeyKvResponse.h
@@ -44,7 +44,7 @@ public:
     /// ScanSkeyKvResponse members
 
     /// <summary>
-    /// 返回的文档数量，为_0_不表示结束。 - 如果filtered_count超过_500_仍无匹配，则返回_0。 -  长度：4
+    /// 返回的文档数量，为0不表示结束。 - 如果filtered_count超过500仍无匹配，则返回0。 -  长度：4
     /// </summary>
 
     int32_t getReturnedCount() const;