+++ kvs/include/huaweicloud/kvs/v1/model/UpdateKvRequestBody.h
@@ -16,10 +16,8 @@
 #include <huaweicloud/kvs/v1/model/Update_fields.h>
 #include <huaweicloud/kvs/v1/model/Condition_expression.h>
 #include <string>
-#include <huaweicloud/kvs/v1/model/Projection_blob.h>
 #include <huaweicloud/kvs/v1/model/Update_blob.h>
 #include <huaweicloud/core/bson/Bson.h>
-#include <vector>
 
 namespace HuaweiCloud {
 namespace Sdk {
@@ -94,15 +92,6 @@ public:
     void setUpdateFields(const Update_fields& value);
 
     /// <summary>
-    /// 对kv_doc有效，返回哪些字段列表，默认全部。 &gt; 字段名或嵌套字段路径用&#39;/&#39;分割 ，e_name为下标。
-    /// </summary>
-
-    std::vector<std::string>& getProjectionFields();
-    bool projectionFieldsIsSet() const;
-    void unsetprojectionFields();
-    void setProjectionFields(const std::vector<std::string>& value);
-
-    /// <summary>
     /// 
     /// </summary>
 
@@ -110,15 +99,6 @@ public:
     bool updateBlobIsSet() const;
     void unsetupdateBlob();
     void setUpdateBlob(const Update_blob& value);
-
-    /// <summary>
-    /// 
-    /// </summary>
-
-    Projection_blob getProjectionBlob() const;
-    bool projectionBlobIsSet() const;
-    void unsetprojectionBlob();
-    void setProjectionBlob(const Projection_blob& value);
 
 
 protected:
@@ -132,12 +112,8 @@ protected:
     bool kvOptionsIsSet_;
     Update_fields updateFields_;
     bool updateFieldsIsSet_;
-    std::vector<std::string> projectionFields_;
-    bool projectionFieldsIsSet_;
     Update_blob updateBlob_;
     bool updateBlobIsSet_;
-    Projection_blob projectionBlob_;
-    bool projectionBlobIsSet_;
 
 };