+++ smn/include/huaweicloud/smn/v2/SmnRegion.h
@@ -110,6 +110,10 @@ public:
             { 
                 "https://smn.my-kualalumpur-1.myhuaweicloud.com"
             });
+        regions["ru-moscow-1"] = Region("ru-moscow-1",
+            { 
+                "https://smn.ru-moscow-1.myhuaweicloud.com"
+            });
         return regions;
     }
 private: